requirements.txt
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.4
pandas==2.1.1
scikit-learn==1.3.2
matplotlib==3.8.0
streamlit==1.28.2
lime==0.2.0.1
fpdf2==2.7.6
tqdm==4.66.1
albumentations==1.3.1
scikit-image==0.21.0
shap==0.44.0
PyYAML==6.0.1
onnxruntime==1.16.1
seaborn==0.13.0
onnx==1.15.0
